#26744 Moved SiteIterator from the 'common' package to a 'site' package.
