+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/api/client/pull/site/SiteFetcher.java
@@ -3,7 +3,6 @@ package com.dotcms.api.client.pull.site;
 import com.dotcms.api.SiteAPI;
 import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.api.client.pull.ContentFetcher;
-import com.dotcms.api.client.pull.common.SiteIterator;
 import com.dotcms.model.ResponseEntityView;
 import com.dotcms.model.site.GetSiteByNameRequest;
 import com.dotcms.model.site.Site;
